Added a toolbar to simple mode and relocated the screenshots
DropStack's simple view now has a toolbar, which can be hidden (in which case it will be represented by a pill you can tap or hover your pointer over to invoke it with a smooth animation) or pinned in settings. It supports all the same keyboard shortcuts as the one in portal view, but adding Ctrl+T to toggle between hiding and pinning the toolbar.

We also relocated the screenshots folder, so it won't be included in released packages anymore, decreasing download sizes.