Working on caching support for portal
DropStack can now save the contents of portal view to an xml file, however, it cannot read them yet. This WIP-feature is turned off by default, resulting in no performance loss out of the box, it needs to be turned on in settings. This might change once it's fully functional. Also made changes so that the search box is not wider than the background rectangle anymore.