Added settings for Pin Visibility with Windows Hello™️ support
Four options, regarding the visibility of pinned files have been added - always, smart, never and protected. Here's what they do:

Always / never - on launch, the bar will be opened or collapsed. If it is collapsed, the app will not load pinned files until it is opened.

Smart: The app will remember if you last had the bar open or closed an will recreate this,

Protected: If chosen, pinned files will now be secured through Windows Hello™️, a.k.a. through your Windows Pin, facial/fingerprint recognition or a security key. All of these need to be set up int he Windows settings menu. Windows handles the security aspect, the app never gets to see the user's face, fingerprint, Pin or security key. To disable Windows hello security, authentication is required. If the user did not set up Windows Hello™️ on their device, the option will be greyed out. If the user enables the setting in the app and then removes Windows Hello in the settings app, the user will need to reinstall the app to reset the setting, something the new options to copy file paths can help with. This feature is intended to prevent kids and other people from immediately getting access to your pinned files, not to withstand planned attacks, since files can still be accessed at the path the user picked for the pinned files anyways.