Simple mode now closes after some actions
Simple mode's window closes after the user completed a drag-out or opened a file. Copying is not affected by this, since the window needs to remain open for the copy to work and cannot be minimized, since it is not visible on the taskbar or the switchers, meaning that it would lead to a memory leak.