Changed app details + fixed bug
Changed DropStack (WinUI)'s name from "DropStack Canary" to "DropStack" and added a new provisory icon, thanks to @totallygeekdom, changed the app's description to drop the early access disclaimer.

Also fixed a bug that would prevent sorting in simple mode to work, as well as applied logic that would allow sorting to work before the app has finished loading (however, it will not update the sorted, please unselect and select the category again to update the list, according to the files that have been loaded).

Co-Authored-By: totallygeekdom <76084843+totallygeekdom@users.noreply.github.com>