Fixed scaling
DropStack now properly scales as you change Windows scaling settings. This includes the great feature of not clipping out of the screen because it falsely calculated where it needed to go, as well as clipping outside of the screen, because it was too tall - both of these are fixed now.