New: Custom performance settings
The settings menu now provides an option to change four performance settings:

- Regular file cap: Amount of files loaded in regular mode, 1000 by default
- Simple file cap: Amount of files loaded in simple mode, 250 by default
- Thumbnail cap: Amount of thumbnails loaded, by default the same number as simple mode
- Thumbnail resolution: Does what it says, default is 64 pixels by 64 pixels, which has been reduced from 256 pixels by 256 pixels, which was very high, given the thumbnail size.

Especially anything with thumbnails in higher resolutions can dramatically impact the app's memory consumption and speed.

There is also a not above that changes made will be reflected after refreshing the app, switching to simple mode or restarting the app.

Additionally, there is a button below to reset the values to their defaults. This can also be done with the launch argument "resetPerformanceSettings". There is also a launch argument that reduces these settings close to their minimum: "lowerPerformanceSettings". Either of these are intended to be used in case the user sets their settings to a very high value, causing the app to crash, as seen when testing a folder with approximately 20.000 files.

Loading less or lower quality thumbnails should also mean being able to load more files.

Also, running the app with launch arguments modifying its settings will now no longer create a window of the app, instead, the app should perform this action in the background and then close, after which the user can reopen the app with the new changes made to the settings being visible.