Introduced caching for portal files
Portal files are now cached to an XML file each time the app finishes loading contents. On next launch, these cached files will immediately show, after which the app will load new files and at them at the top, loading file thumbnails and removing files that do not exist anymore. We also applied this to dragging files out of normal view, where it won't refresh the entire thing anymore, but instead check if the file is still there or not.

Furthermore, because simple mode is responsible for indexing as well, the setting to control how many items are loaded in now controls both simple and normal mode. We also renamed the loaded thumbnails setting, to also align with this and removed its reference to the simple mode setting.

Furthermore, setting the loaded files amount to the minimum using the console command no longer sets it below the minimum allowed value of 20.

About this limitation, DropStack will always load the appropriate number of images that you set in the settings, however, it will load the set number of loaded files, plus new files. This is because the trimming is now done while saving and loading files, meaning that old files are still there, but they are not being updated anymore - they also do not show a thumbnail. Furthermore, they will not be included in the next caching and thus will not show up when the app launches next time anymore. Refreshing also removes them. They are still being deleted in case they are non-existent, though.