Caching for pinned files in normal mode
Caching now works for pinned files in normal mode - in simple mode, this is not as critical, since files are not loaded in at launch and are far fewer in number. They are still being cached after they are loaded. If you have pins protected with Windows Hello, they will not load until you authenticated.