Showing checkboxes for multi-select
Regular file list view now has the option to show checkboxes next to its files to make multi select easier. As of now, it also supports using the Windows Explorer setting, however, it does not support detecting whether the device is in tablet mode - the event handler does not work in either case.