Disconnect button now says relaunch, not quit
The disconnect folder button confirmation button now mentions the app relaunching instead of quitting, since the app relaunches and doesn't quit.

The text changed from "Yes, disconnect and quit!" to "Yes, disconnect and relaunch!"

Also removed the event handler that was supposed to detect when a keyboard was being detached, since it did not work anyway.