Reworked Search Experience
Settings now include search leniency settings that are preserved between sessions. This value controls how lenient DropStack's file search should be in terms of typos. Basically, if this value is set to three (which is the default), DropStack will still show files in search that can be found by changing three characters in the search query. If the searched text is shorter or equal to the leniency setting, the leniency will be reduced to the search query length minus one.

Also added a short delay to search so it only refreshes when you stopped typing, which improves performance.